Standardize references in SeMRA sources (#36)
This PR switches the generic `curies.Reference` for a `pyobo.Reference`,
which checks that the prefix and identifier are standardized against the
Bioregistry. This applies only to the `semra.sources` modules

In theory, this should catch the issue in #33. I added a CLI to run just
against semra sources `python -m semra.sources` to check this (running
now)