Add PURL to mapping set to align with SSSOM's `mapping_set_id` (#65)
Related to #60 

This PR adds an (optional) `purl` field to the `MappingSet` class to
better align with SSSOM's `mapping_set_id` field. For input SSSOM files,
this isn't required, but this is useful when outputting SSSOM.

Warning: this breaks all previous mapping set hashes, since it now adds
the ID into it.