Add tests for web implementation (#75)
This PR abstracts the neo4j client into a base client that can be mocked
in tests. This PR then adds tests for the fastapi and flask endpoints.

Future work:

1. More detailed testing of summary components
2. How to test the cytoscape component output? Unify the cytoscape graph
model with the other code that constructs these
3. Testing of the Neo4j client itself (requires mocking neo4j / having a
test instance)

At least this one catches a few issues in the templates