Fix connected component query for graph display (#99)
This PR fixes two issues with the way queries are done to get connected
components in the web app.

Before

<img width="1338" height="970" alt="image"
src="https://github.com/user-attachments/assets/3137b768-3fcd-4490-b1aa-d8f8218be3b8"
/>


After

<img width="1387" height="1089" alt="image"
src="https://github.com/user-attachments/assets/f2b01bd0-b644-4a60-97b3-fae29b97ba50"
/>

To elaborate:
- First issue
- The original query's intent was to filter to only mappings with
primary and secondary evidence to avoid redundant mappings in the graph
that are a result of inversion or chaining.
- However, this has the unintended effect of also filtering out mappings
that are inferred from `hasDbXref` to `exactMatch` and are considered
tertiary.
- This results in the graph being incomplete and not containing all the
nodes that are actually part of the connected component (i.e., the terms
that show up in the table on the same page).
- To solve this, the query is made more explicit to specifically
eliminate `semapv:MappingInversion` and `semapv:MappingChaining` which
results in recovering the intended graph over the nodes in question.
  
- Second issue
- The list of entities in the table above the graph comes from
`get_exact_matches` which looks _only_ at `skos:exactMatch` as a
relationship. In contrast, by default, `get_connected_component` uses
`self._rel_q` by default which also includes `oboinowl:hasDbXref`. This
has no effect when we're looking at a processed database, however, it is
a major issue for a raw database where this difference results in a
completely different graph.
- To solve this, I also added an additional argument to
`get_connected_component` and any other necessary places to pass in a
custom relation constraint which, in the web app is set to just
`skos:exactMatch`.
- Overall, this ensures that the table of entities and the Cytoscape
graph are fully consistent.